add search by tag functionality
