+++ views/partials/courseCard.handlebars
@@ -2,9 +2,7 @@
     <div class="card mx-auto" style="width: 18rem">
     
       <iframe src="{{course_link}}" 
-        allow="accelerometer; autoplay; clipboard-write; encrypted-media; 
-          gyroscope; picture-in-picture" 
-        allowfullscreen>
+        allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen>
       </iframe>
 
       <div class="card-body">
